% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo.R
\name{denovo.sensi}
\alias{denovo.sensi}
\title{De novo discovery of effect modification with a sensitivity analysis.}
\usage{
denovo.sensi(Data, Gamma.vec, split.ratio = 0.25, total.significance = 0.05,
  gamma = 1e-04)
}
\arguments{
\item{Data}{a total sample of matched pairs.}

\item{Gamma.vec}{a vector of sensitivity parameter \eqn{\Gamma}.}

\item{split.ratio}{set a splitting ratio. 0.25 means using 25% of data as a training sample.}

\item{total.significance}{a total significance level.}

\item{gamma}{a significance level that is used for estimating the confidence interval.}
}
\value{
A result matrix with deviates is reported. The `Max` column represents the maximum value of deviates.
The `kappa` column represents the critical value. If `Max` is greater than 'kappa', we reject the null hypothesis.
}
\description{
De novo discovery of effect modification with a sensitivity analysis.
}
