% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo.R
\name{denovo.test}
\alias{denovo.test}
\title{Conduct a hypothesis test under no unmeasured confounder assumption.}
\usage{
denovo.test(Data, test.index, tree, total.significance = 0.05, gamma = 0.01)
}
\arguments{
\item{Data}{a total sample of matched pairs.}

\item{test.index}{an index set that indicates a test sample in the total sample.}

\item{tree}{an effect modification tree that needs to be tested.}

\item{total.significance}{a total significance level.}

\item{gamma}{a significance level that is used for estimating the confidence interval.}
}
\value{
A result matrix with deviates is reported. The `Max` column represents the maximum value of deviates.
The `kappa` column represents the critical value.
}
\description{
Conduct a hypothesis test under no unmeasured confounder assumption.
}
