% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo.R
\name{denovo.test.sensi}
\alias{denovo.test.sensi}
\title{Conduct a sensitivity analysis}
\usage{
denovo.test.sensi(Data, test.index, tree, Gamma.vec,
  total.significance = 0.05, gamma = 1e-04)
}
\arguments{
\item{Data}{a total sample of matched pairs.}

\item{test.index}{an index set that indicates a test sample in the total sample.}

\item{tree}{an effect modification tree that needs to be tested.}

\item{Gamma.vec}{a vector of sensitivity parameter \eqn{\Gamma}.}

\item{total.significance}{a total significance level.}

\item{gamma}{a significance level that is used for estimating the confidence interval.}
}
\value{
A result matrix with deviates is reported. The `Max` column represents the maximum value of deviates.
The `kappa` column represents the critical value. If `Max` is greater than 'kappa', we reject the null hypothesis.
}
\description{
Conduct a sensitivity analysis
}
