% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo.R
\name{denovo.training}
\alias{denovo.training}
\title{Find an effect modification structure from data.}
\usage{
denovo.training(Data, split.ratio = 0.25)
}
\arguments{
\item{Data}{input matched pairs. The first column must contain outcomes for treated subjects and the second column must contain outcomes for control subjects. The remaining columns contain shared covariates.}

\item{split.ratio}{set a splitting ratio. 0.25 means using 25\% of data as a training sample.}
}
\value{
A list with the elements:
\item{training.index}{an index set for indicating a training sample that is randomly chosen.}
\item{test.index}{an index set for indicating a test sample.}
\item{tree}{discovered effect modification tree structure.}
}
\description{
Find an effect modification structure from data.
}
