% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get.tau.vector.wilcox}
\alias{get.tau.vector.wilcox}
\title{Estimation of the confidence interval for the population mean by inverting the Wilcoxon signed rank sum test.}
\usage{
get.tau.vector.wilcox(data, gamma = 0.001, grid.size = 21)
}
\arguments{
\item{data}{input a test sample.}

\item{gamma}{set a significance level for the confidence interval estimation.}

\item{grid.size}{the number of values within the confidence interval.}
}
\value{
The \eqn{100(1-\gamma)}% confidence interval for the population mean is estimated.
}
\description{
Estimation of the confidence interval for the population mean by inverting the Wilcoxon signed rank sum test.
}
