% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{tree.stru}
\alias{tree.stru}
\title{Combine a tree structure with a test sample}
\usage{
tree.stru(tree, x)
}
\arguments{
\item{tree}{a tree object}

\item{x}{a test sample with the same data structure that is used for creating `tree`.}
}
\value{
a list with the elements
\item{C}{the matrix representation of a tree.}
\item{new.x}{a new test sample by adding a vector of subgroup memberships.}
}
\description{
The fuction combines a tree object (obtained from a training sample) with a test sample.
}
