% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{wilcoxSenMoments}
\alias{wilcoxSenMoments}
\title{The Wilcoxon's signed rank sum test in a sensitivity analysis.}
\usage{
wilcoxSenMoments(N, gamma)
}
\arguments{
\item{N}{the number of matched pairs}

\item{gamma}{the value of sensitivity parameter \eqn{\Gamma}}
}
\value{
\item{expect.upper}{Expectation of the upper bound.}
\item{expect.lower}{Expectation of the lower bound.}
\item{var}{Variance of both upper and lower bounds.}
}
\description{
The Wilcoxon's signed rank sum test in a sensitivity analysis.
}
